Added parameter verifications, small optimization.

+++ headers/Bloomfilter.hpp
@@ -7,11 +7,11 @@
 using namespace std;
 
 class Bloomfilter {
-    uint64_t size;
-    uint32_t n_hashes;
+    uint32_t size;
+    uint8_t n_hashes;
     vector<bool> bit_set;
 public:
-    Bloomfilter(uint64_t n, uint32_t hashes);
+    Bloomfilter(uint32_t n, uint8_t hashes);
     void add_value(uint64_t x);
     bool is_present(uint64_t x);
 };
